move telemetry calculation to module
